+++ web/src/beta/features/Editor/tabs/story/StoryPageIndicator/index.stories.tsx
