+++ web/src/beta/features/Editor/tabs/story/StoryPageIndicator/index.tsx
