+++ web/src/beta/features/Editor/tabs/story/StoryPanel/index.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 
-import StoryPageIndicator from "@reearth/beta/features/Editor/tabs/Storytelling/StoryPageIndicator";
+import StoryPageIndicator from "@reearth/beta/features/Editor/tabs/story/StoryPageIndicator";
 import { styled } from "@reearth/services/theme";
 
 type Props = {};
