+++ web/src/beta/features/Editor/tabs/story/StorySidePanelAction/index.stories.tsx
