+++ web/src/beta/features/Editor/tabs/story/StorySidePanelItem/index.stories.tsx
